#38
- preset/isomorphic: утилиты разнесены по отдельным файлам (major)
- добавлены тесты config и utils/axios (patch)
