+++ src/preset/isomorphic/providers/index.ts
@@ -3,7 +3,8 @@ import { BaseConfig } from '../../../config/types';
 import { Resolve } from '../../../di';
 import { LogMiddlewareHandlerInit } from '../../../utils/axios/middleware/log';
 import { KnownToken } from '../../../tokens';
-import { AxiosLogging, SagaLogging } from '../utils';
+import { AxiosLogging } from '../utils/axios-logging';
+import { SagaLogging } from '../utils/saga-logging';
 import createSagaMiddleware, { SagaMiddleware } from 'redux-saga';
 import { applyMiddleware, configureFetch } from '../../../http';
 import { CreateAxiosDefaults } from 'axios';
